+++ src/ergotest/test_result.ts
@@ -102,6 +102,8 @@ export abstract class TestResult {
 	/**
 	 * Create a TestResult for a test that passed.
 	 * @param {string|string[]} name The name of the test. Can be a list of names.
+	 * @param {TestCaseResult[]} [options.beforeEach] The beforeEach() blocks for this test.
+	 * @param {TestCaseResult[]} [options.afterEach] The afterEach() blocks for this test.
 	 * @param {string} [options.filename] The file that contained this test (optional).
 	 * @param {TestMarkValue} [options.mark] Whether this test was marked with `.skip`, `.only`, or nothing.
 	 * @returns {TestCaseResult} The result.
@@ -109,9 +111,13 @@ export abstract class TestResult {
 	static pass(
 		name: string | string[],
 		{
+			beforeEach,
+			afterEach,
 			filename,
 			mark,
 		}: {
+			beforeEach?: TestCaseResult[],
+			afterEach?: TestCaseResult[],
 			filename?: string,
 			mark?: TestMarkValue,
 		} = {}
@@ -119,13 +125,15 @@ export abstract class TestResult {
 		ensure.signature(arguments, [
 			[ String, Array ],
 			[ undefined, {
+				beforeEach: [ undefined, Array ],
+				afterEach: [ undefined, Array ],
 				filename: [ undefined, String ],
 				mark: [ undefined, String ]
 			}],
 		]);
 
 		if (!Array.isArray(name)) name = [ name ];
-		return new TestCaseResult({ name, status: TestStatus.pass, filename, mark });
+		return new TestCaseResult({ name, status: TestStatus.pass, beforeEach, afterEach, filename, mark });
 	}
 
 	/**
@@ -134,6 +142,8 @@ export abstract class TestResult {
 	 * @param {unknown} error The error that occurred.
 	 * @param {(name: string, error: unknown, mark: TestMarkValue, filename?: string) => unknown} [options.renderError]
 	 *   The function to use to render the error into a string (defaults to {@link renderError})
+ *   @param {TestCaseResult[]} [options.beforeEach] The beforeEach() blocks for this test.
+	 * @param {TestCaseResult[]} [options.afterEach] The afterEach() blocks for this test.
 	 * @param {string} [options.filename] The file that contained this test (optional).
 	 * @param {TestMarkValue} [options.mark] Whether this test was marked with `.skip`, `.only`, or nothing.
 	 *   function will be called and the results put into {@link errorRender}.
@@ -144,10 +154,14 @@ export abstract class TestResult {
 		error: unknown,
 		{
 			renderError = renderErrorFn,
+			beforeEach,
+			afterEach,
 			filename,
 			mark,
 		}: {
 			renderError?: RenderErrorFn
+			beforeEach?: TestCaseResult[],
+			afterEach?: TestCaseResult[],
 			filename?: string,
 			mark?: TestMarkValue,
 		} = {},
@@ -157,6 +171,8 @@ export abstract class TestResult {
 			ensure.ANY_TYPE,
 			[ undefined, {
 				renderError: [ undefined, Function ],
+				beforeEach: [ undefined, Array ],
+				afterEach: [ undefined, Array ],
 				filename: [ undefined, String ],
 				mark: [ undefined, String ],
 			}],
@@ -171,12 +187,16 @@ export abstract class TestResult {
 
 		const errorRender = renderError(name, error, mark ?? TestMark.none, filename);
 
-		return new TestCaseResult({ name, status: TestStatus.fail, errorMessage, errorRender, filename, mark });
+		return new TestCaseResult(
+			{ name, status: TestStatus.fail, errorMessage, errorRender, beforeEach, afterEach, filename, mark }
+		);
 	}
 
 	/**
 	 * Create a TestResult for a test that was skipped.
 	 * @param {string|string[]} name The name of the test. Can be a list of names.
+	 * @param {TestCaseResult[]} [options.beforeEach] The beforeEach() blocks for this test.
+	 * @param {TestCaseResult[]} [options.afterEach] The afterEach() blocks for this test.
 	 * @param {string} [options.filename] The file that contained this test (optional).
 	 * @param {TestMarkValue} [options.mark] Whether this test was marked with `.skip`, `.only`, or nothing.
 	 * @returns {TestCaseResult} The result.
@@ -184,9 +204,13 @@ export abstract class TestResult {
 	static skip(
 		name: string | string[],
 		{
+			beforeEach,
+			afterEach,
 			filename,
 			mark,
 		}: {
+			beforeEach?: TestCaseResult[],
+			afterEach?: TestCaseResult[],
 			filename?: string,
 			mark?: TestMarkValue,
 		} = {}
@@ -194,19 +218,23 @@ export abstract class TestResult {
 		ensure.signature(arguments, [
 			[ String, Array ],
 			[ undefined, {
+				beforeEach: [ undefined, Array ],
+				afterEach: [ undefined, Array ],
 				filename: [ undefined, String ],
 				mark: [ undefined, String ]
 			}],
 		]);
 
 		if (!Array.isArray(name)) name = [ name ];
-		return new TestCaseResult({ name, status: TestStatus.skip, filename, mark });
+		return new TestCaseResult({ name, status: TestStatus.skip, beforeEach, afterEach, filename, mark });
 	}
 
 	/**
 	 * Create a TestResult for a test that timed out.
 	 * @param {string|string[]} name The name of the test. Can be a list of names.
 	 * @param {number} timeout The length of the timeout.
+	 * @param {TestCaseResult[]} [options.beforeEach] The beforeEach() blocks for this test.
+	 * @param {TestCaseResult[]} [options.afterEach] The afterEach() blocks for this test.
 	 * @param {string} [options.filename] The file that contained this test (optional).
 	 * @param {TestMarkValue} [options.mark] Whether this test was marked with `.skip`, `.only`, or nothing.
 	 * @returns {TestCaseResult} The result.
@@ -215,9 +243,13 @@ export abstract class TestResult {
 		name: string | string[],
 		timeout: number,
 		{
+			beforeEach,
+			afterEach,
 			filename,
 			mark,
 		}: {
+			beforeEach?: TestCaseResult[],
+			afterEach?: TestCaseResult[],
 			filename?: string,
 			mark?: TestMarkValue,
 		} = {},
@@ -226,13 +258,15 @@ export abstract class TestResult {
 			[ String, Array ],
 			Number,
 			[ undefined, {
+				beforeEach: [ undefined, Array ],
+				afterEach: [ undefined, Array ],
 				filename: [ undefined, String ],
 				mark: [ undefined, String ]
 			}],
 		]);
 		
 		if (!Array.isArray(name)) name = [ name ];
-		return new TestCaseResult({ name, status: TestStatus.timeout, timeout, filename, mark });
+		return new TestCaseResult({ name, status: TestStatus.timeout, timeout, beforeEach, afterEach, filename, mark });
 	}
 
 	/**
@@ -371,16 +405,16 @@ export class TestSuiteResult extends TestResult {
 	}
 
 	/**
-	 * @returns { TestResult[] } The beforeAll() blocks ran for this suite.
+	 * @returns { TestCaseResult[] } The beforeAll() blocks for this suite.
 	 */
-	get beforeAll(): TestResult[] {
+	get beforeAll(): TestCaseResult[] {
 		return this._beforeAll;
 	}
 
 	/**
-	 * @returns { TestResult[] } The afterAll() blocks ran for this suite.
+	 * @returns { TestCaseResult[] } The afterAll() blocks for this suite.
 	 */
-	get afterAll(): TestResult[] {
+	get afterAll(): TestCaseResult[] {
 		return this._afterAll;
 	}
 
@@ -594,6 +628,8 @@ export class TestCaseResult extends TestResult {
 
 	private _name: string[];
 	private _filename?: string;
+	private _beforeEach: TestCaseResult[];
+	private _afterEach: TestCaseResult[];
 	private _status: TestStatusValue;
 	private _mark: TestMarkValue;
 	private _errorMessage?: string;
@@ -608,6 +644,8 @@ export class TestCaseResult extends TestResult {
 			errorMessage,
 			errorRender,
 			timeout,
+			beforeEach = [],
+			afterEach = [],
 			filename,
 			mark,
 		}: {
@@ -616,6 +654,8 @@ export class TestCaseResult extends TestResult {
 			errorMessage?: string,
 			errorRender?: unknown,
 			timeout?: number,
+			beforeEach?: TestCaseResult[],
+			afterEach?: TestCaseResult[],
 			filename?: string,
 			mark?: TestMarkValue
 		},
@@ -623,6 +663,8 @@ export class TestCaseResult extends TestResult {
 		super();
 		this._name = name;
 		this._filename = filename;
+		this._beforeEach = beforeEach;
+		this._afterEach = afterEach;
 		this._status = status;
 		this._mark = mark ?? TestMark.none;
 		this._errorMessage = errorMessage;
@@ -650,6 +692,20 @@ export class TestCaseResult extends TestResult {
 	 */
 	get mark(): TestMarkValue {
 		return this._mark;
+	}
+
+	/**
+	 * @returns { TestCaseResult[] } The beforeEach() blocks for this test.
+	 */
+	get beforeEach(): TestCaseResult[] {
+		return this._beforeEach;
+	}
+
+	/**
+	 * @returns { TestCaseResult[] } The afterEach() blocks for this test.
+	 */
+	get afterEach(): TestCaseResult[] {
+		return this._afterEach;
 	}
 
 	/**