+++ CHANGELOG.md
@@ -20,6 +20,7 @@ To Document:
 * TestResult.skip signature change
 * TestResult.timeout signature change
 * Add TestSuiteResult.beforeAll & afterAll
+* Add TestCaseResult.beforeEach & afterEach
 * TestSuiteResult.children --> TestSuiteResult.tests rename
 * Change in behavior: beforeAll() and afterAll() now notify upon completion (change name of onTestCaseComplete? To onProgress maybe?)
 * Change in behavior: beforeAll() and afterAll() failures no longer result in failing TestCaseResults. Instead, they show up in TestSuiteResult.beforeAll or .afterAll